Вынос констант, применение сортировки, и ограничения по количеству на полученные комментарии.
