+++ src/components/review-list/review-list.tsx
@@ -2,16 +2,17 @@ import Review from '../review/review';
 import { CommentForOffer} from '../../types';
 
 type CommentListProps = {
-  commentsByOffer : CommentForOffer[];
+  countAllComments : number;
+  commentsByOfferSorted : CommentForOffer[];
 }
 
-export default function ReviewList({ commentsByOffer}: CommentListProps): JSX.Element {
+export default function ReviewList({countAllComments,commentsByOfferSorted}: CommentListProps): JSX.Element {
 
   return (
     <>
-      <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{commentsByOffer.length}</span></h2>
+      <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{countAllComments}</span></h2>
       <ul className="reviews__list" >
-        {commentsByOffer.map((comment) => (
+        {commentsByOfferSorted.map((comment) => (
           <Review comment={comment} key={comment.id} />
         ))}
       </ul>
